refactor(gsd): unify skill loading and wire skillFilter
Centralize skill discovery on the ResourceLoader facade, apply manifest
filtering to the system prompt catalog, collapse duplicate discovery
injection, align precedence docs with PackageManager behavior, and split
oversized skill modules for maintainability.
