+++ src/resources/extensions/gsd/auto.ts
@@ -135,6 +135,7 @@ import {
   captureAvailableSkills,
   resetSkillTelemetry,
 } from "./skill-telemetry.js";
+import { getInstalledSkillNames } from "./skills.js";
 import { getRtkSessionSavings } from "../shared/rtk-session-stats.js";
 import { deactivateGSD } from "../shared/gsd-phase-state.js";
 import {
@@ -2562,7 +2563,7 @@ function buildLoopDeps(pi: ExtensionAPI): LoopDeps {
     autoCommitUnit,
     recordOutcome,
     writeLock,
-    captureAvailableSkills,
+    captureAvailableSkills: () => captureAvailableSkills(getInstalledSkillNames()),
     ensurePreconditions,
     updateSliceProgressCache,
